+++ packages/ui5-application-writer/test/options.test.ts
@@ -105,6 +105,10 @@ describe('UI5 templates', () => {
         const projectDir = join(outputDir, 'testapp_options');
         const fs = await generate(projectDir, {
             ...baseAppConfig,
+            app: {
+                ...baseAppConfig.app,
+                flpAction: 'display'
+            },
             appOptions: {
                 sapux: true
             },
@@ -132,6 +136,28 @@ describe('UI5 templates', () => {
                     typescript: true,
                     sapux: true,
                     npmPackageConsumption: true
+                }
+            })
+        );
+        expect(fs.dump(projectDir)).toMatchSnapshot();
+        return new Promise((resolve) => {
+            // write out the files for debugging
+            if (debug) {
+                fs.commit(resolve);
+            } else {
+                resolve(true);
+            }
+        });
+    });
+
+    it('option: `sapux and use virtual endpoints`', async () => {
+        const projectDir = join(outputDir, 'testapp_typescript');
+        const fs = await generate(
+            projectDir,
+            Object.assign(baseAppConfig, {
+                appOptions: {
+                    sapux: true,
+                    useVirtualPreviewEndpoints: true
                 }
             })
         );