+++ packages/abap-deploy-config-writer/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @sap-ux/abap-deploy-config-writer
 
+## 0.0.106
+
+### Patch Changes
+
+-   Updated dependencies [7590bc3]
+    -   @sap-ux/ui5-config@0.27.2
+    -   @sap-ux/project-access@1.29.21
+    -   @sap-ux/system-access@0.5.36
+
+## 0.0.105
+
+### Patch Changes
+
+-   Updated dependencies [294bbe3]
+    -   @sap-ux/project-access@1.29.20
+    -   @sap-ux/ui5-config@0.27.1
+    -   @sap-ux/system-access@0.5.36
+
+## 0.0.104
+
+### Patch Changes
+
+-   Updated dependencies [1ca4004]
+    -   @sap-ux/ui5-config@0.27.0
+    -   @sap-ux/project-access@1.29.19
+    -   @sap-ux/system-access@0.5.36
+
 ## 0.0.103
 
 ### Patch Changes