+++ packages/cf-deploy-config-inquirer/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @sap-ux/cf-deploy-config-inquirer
 
+## 0.2.33
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.6.39
+
+## 0.2.32
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.6.38
+
+## 0.2.31
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.6.37
+
 ## 0.2.30
 
 ### Patch Changes